+++ src/Settings/components/reminders/reminders.jsx
@@ -15,25 +15,21 @@ import { getBaniList } from "@database";
 import { nightModeStyles, iconNightColor } from "../../styles";
 import { ListItemComponent, BottomSheetComponent } from "../comon";
 import setDefaultReminders from "./ReminderOptions/utils";
+import { getReminderSound } from "../comon/strings";
 
 const RemindersComponent = ({ navigation }) => {
-  const REMINDER_SOUNDS = [
-    {
-      key: "default",
-      title: STRINGS.default,
-    },
-    { key: "wake_up_jap.mp3", title: STRINGS.wake_up_jap },
-    { key: "waheguru_soul.mp3", title: STRINGS.waheguru_soul },
-  ];
+  const REMINDER_SOUNDS = getReminderSound(STRINGS);
   const isNightMode = useSelector((state) => state.isNightMode);
   const isReminders = useSelector((state) => state.isReminders);
   const reminderSound = useSelector((state) => state.reminderSound);
   const transliterationLanguage = useSelector((state) => state.transliterationLanguage);
   const [isReminderSound, toggleReminderSound] = useState(false);
+
   const dispatch = useDispatch();
   const { navigate } = navigation;
   const { containerNightStyles, textNightStyle } = nightModeStyles(isNightMode);
   const iconColor = iconNightColor(isNightMode);
+
   const redirectToSettings = async () => {
     Alert.alert(STRINGS.permissionTitle, STRINGS.premissionDescription, [
       {
